pyproject.toml
[project]
name = "actor-claude-sdk-playwright"
version = "0.1.0"
description = "Apify actor with Claude Agent SDK and Playwright MCP for browser automation"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "apify>=2.2.1,<3",
    "claude-agent-sdk>=0.1.0",
    "python-dotenv>=1.0.0",
    "langfuse>=2.0.0",
    "pillow>=10.0.0",
    "pydantic>=2.9,<2.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "ruff>=0.6",
]

[tool.apify]
vendor = "internal"
actor = "claude-sdk-playwright"

[tool.ruff]
line-length = 120

[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
